rename files in camelcase
