+++ tezos-sig-gating-contracts/scripts/manual/deployNftminter.ts
