+++ tezos-sig-gating-contracts/scripts/manual/localDeployNftminter.ts
