+++ tezos-sig-gating-contracts/scripts/manual/localDisplayStorage.ts
