+++ tezos-sig-gating-contracts/scripts/manual/localMintOffchain.ts
