+++ tezos-sig-gating-contracts/scripts/manual/localSign.ts
