+++ tezos-sig-gating-contracts/scripts/manual/mintOffchain.ts
